Move test for metadata extraction at new dataset version to `TestExtractDsMeta`
