--- datalad_registry/tests/test_tasks/test_extract_meta.py
@@ -1,84 +1,0 @@
-from pathlib import Path
-
-from datalad import api as dl
-from datalad.distribution.dataset import require_dataset
-import pytest
-
-from datalad_registry.tasks import ExtractMetaStatus, extract_ds_meta
-
-_TEST_REPO_URL = "https://github.com/datalad/testrepo--minimalds.git"
-_TEST_REPO_TAG = "0.1.0"
-_TEST_REPO_COMMIT_HEXSHA = "ac9ba85cf1e8a004e7c24ebb6b5cd861d53e3998"
-_BASIC_EXTRACTOR = "metalad_core"
-
-
-@pytest.fixture
-def test_repo_url_id(flask_app) -> int:
-    from datalad_registry.models import URL, db
-
-    url = URL(url=_TEST_REPO_URL, head_describe=_TEST_REPO_TAG)
-
-    with flask_app.app_context():
-        db.session.add(url)
-        db.session.commit()
-        return url.id
-
-
-@pytest.fixture(scope="session")
-def test_repo_path(tmp_path_factory):
-    test_repo_path_ = tmp_path_factory.mktemp("test_repo")
-    dl.clone(
-        _TEST_REPO_URL,
-        path=test_repo_path_,
-        git_clone_opts=[f"--branch={_TEST_REPO_TAG}"],
-    )
-    return test_repo_path_
-
-
-class TestExtractMeta:
-    def test_new_dataset_version(self, flask_app, processed_ds_urls):
-        """
-        Test extraction of metadata for a dataset at a new version after the extraction
-        of metadata for the same dataset at an older version
-
-        The extraction should return ExtractMetaStatus.SUCCEEDED, and there should be
-        only one piece of metadata saved to database, the latest one
-        """
-        from datalad_registry.models import URL, URLMetadata, db
-
-        test_repo_url_id = processed_ds_urls[0]
-
-        # Fetch test repo cache path
-        with flask_app.app_context():
-            test_repo_path = str(
-                Path(flask_app.config["DATALAD_REGISTRY_DATASET_CACHE"])
-                / db.session.execute(
-                    db.select(URL.cache_path).where(URL.id == test_repo_url_id)
-                ).scalar_one()
-            )
-
-        test_ds = require_dataset(test_repo_path, check_installed=True, purpose="test")
-
-        # Set dataset one commit before the reference tag for test
-        test_ds.repo.call_git(["checkout", f"{_TEST_REPO_TAG}^"])
-
-        with flask_app.app_context():
-            # Extract metadata for the dataset at the older version
-            ret = extract_ds_meta(test_repo_url_id, _BASIC_EXTRACTOR)
-            assert ret == ExtractMetaStatus.SUCCEEDED
-
-            # There should be only one piece of metadata saved to database
-            dated_metadata = db.session.execute(db.select(URLMetadata)).scalar_one()
-
-            # Set dataset to the reference tag for test
-            test_ds.repo.call_git(["checkout", _TEST_REPO_TAG])
-
-            # Extract metadata for the dataset at the new version
-            ret = extract_ds_meta(test_repo_url_id, _BASIC_EXTRACTOR)
-            assert ret == ExtractMetaStatus.SUCCEEDED
-
-            # There should be only one piece of metadata saved to database
-            current_metadata = db.session.execute(db.select(URLMetadata)).scalar_one()
-
-            assert current_metadata.dataset_version != dated_metadata.dataset_version
-            assert current_metadata.dataset_version == _TEST_REPO_COMMIT_HEXSHA
